+++ tokenManager.js
@@ -22,15 +22,31 @@ let repository = new Repository('Tokens');
 let tokenLifeDuration = serverVariables.get("main.token.lifeDuration");
 
 class TokenManager{
-    static create(email) {
+    static create(user) {
         let token = {   Id: 0, 
-                        Access_token: makeToken(email), 
-                        Expires_in: utilities.nowInSeconds() + tokenLifeDuration
+                        Access_token: makeToken(user.Email), 
+                        Expires_in: utilities.nowInSeconds() + tokenLifeDuration,
+                        UserId: user.Id,
+                        Username: user.Name
                     };
+        console.log("User " + token.Username + " logged in");
         repository.add(token);
         return token;
     }
-    static cleanTokens() {
+    static logout(userId) {
+        let tokens = repository.getAll();
+        let index = 0;
+        let indexToDelete = [];
+        for(let token of tokens) {
+            if (token.UserId == userId) {
+                console.log("User " + token.Username + " logged out");
+                indexToDelete.push(index);
+            }
+            index ++;
+        }
+        repository.removeByIndex(indexToDelete);
+    }
+    static clean() {
         let tokens = repository.getAll();
         let now = utilities.nowInSeconds();
         let index = 0;
@@ -42,7 +58,8 @@ class TokenManager{
             }
             index ++;
         }
-        repository.removeByIndex(indexToDelete);
+        if (index > 0)
+            repository.removeByIndex(indexToDelete);
     }
     static find(access_token) {
         let token = repository.findByField('Access_token', access_token);
@@ -66,5 +83,5 @@ class TokenManager{
 
 // periodic cleaning of expired tokens
 console.log("Periodic tokens cache cleaning process started...");
-setInterval(TokenManager.cleanTokens, tokenLifeDuration * 1000);
+setInterval(TokenManager.clean, tokenLifeDuration * 1000);
 module.exports = TokenManager;
